requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
appdirs==1.4.4
    # via python-escpos
argcomplete==3.1.1
    # via python-escpos
argparse==1.4.0
    # via python-escpos
couchdb==1.2
    # via librocco-receipt-printer (setup.py)
pillow==10.0.0
    # via python-escpos
pypng==0.20220715.0
    # via qrcode
pyserial==3.5
    # via python-escpos
python-escpos==2.2.0
    # via librocco-receipt-printer (setup.py)
pyusb==1.2.1
    # via python-escpos
pyyaml==6.0.1
    # via python-escpos
qrcode==7.4.2
    # via python-escpos
six==1.16.0
    # via python-escpos
typing-extensions==4.7.1
    # via qrcode
